+++ src/controllers/AuthorizeController.php
@@ -15,6 +15,7 @@ use Exception;
 use League\OAuth2\Client\Provider\Exception\IdentityProviderException;
 use ReflectionException;
 use venveo\oauthclient\base\Provider;
+use venveo\oauthclient\events\AuthorizationUrlEvent;
 use venveo\oauthclient\models\App as AppModel;
 use venveo\oauthclient\models\Token as TokenModel;
 use venveo\oauthclient\Plugin;
@@ -54,6 +55,7 @@ class AuthorizeController extends Controller
             return null;
         }
 
+        $context = Craft::$app->request->getParam('context');
         $error = Craft::$app->request->getParam('error');
         $code = Craft::$app->request->getParam('code');
         $state = Craft::$app->request->getParam('state');
@@ -72,11 +74,11 @@ class AuthorizeController extends Controller
         // Begin auth process
         if (empty($code)) {
             $state = $this->getRandomState();
-            $url = $provider->getAuthorizeURL(['state' => $state]);
             Craft::$app->session->set(self::STATE_SESSION_KEY, $state);
-            return Craft::$app->response->redirect($url);
 
-            // Invalid state
+            $url = Plugin::$plugin->apps->getAuthorizationUrlForApp($app, $state, $context);
+
+            return Craft::$app->response->redirect($url);
         }
 
         if (empty($state) || Craft::$app->session->get(self::STATE_SESSION_KEY) !== $state) {